blocks falling at correct time
the game is now somewhat working, all the blocks fall at the right time following the melody array. the sounds play when the keys are pressed and repeat notes are now wokring. Next issue is that there is a delay once the block is pressed before the users are able to press the next key.